+++ crates/authly-common/CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - decouple policies from services, introduce `domain` concept into document model.
 - improved implementation of policy engine with new logical model for policy combinations.
 - `entity-attribute-binding` renamed to `entity-attribute-assignment`. Now accepts only one parameter for identifying the entity.
+- Protobuf definitions changed into using `authly.` package prefix.
 
 ## [0.0.5] - 2025-01-30
 ### Changed